More coloring
I'm working my way toward visualizing
all tools (not just those that provide
cross-check results).  Part of this is on
the fmi-data side (in generating the
report) and part of it is here in the
rendering.

With all tools, things are getting pretty
"crowded".  As such, I view for the
selected tool will need to be either a
popover or some kind of router based
navigation in order to fit everything on
one screen.

The current state of things is a bit of
a work in progress.